Added new screen and made design more consistent
Added a settings screen, which takes the place of the help screen's button
while the help screen is relegated to the "?" image. This will allow the user
to have he useful buttons all in a row at the bottoms, while the help menu
is separate and more visible for users who are unfamiliar with the interface.

Made design more consistent by updating the navigation buttons on the settings
and help screens, making it a more seamless experience when transitioning between
screens. This consists of a light gray border at the bottom that wraps around
a darker gray button. This will make each screen seem like they are more from
the same product instead of disjointed sections.